fix(handler,rbac,postgres): round-28 cascade updated_at + typed envelope + int64 TTL + binding cap warn
Devin Review on PR #42 (commit 4275508) round-28 flagged five items
spanning two flagged-as-bug findings and three info-level long-term
hardening opportunities. All five addressed; one informational ack
on the Go-version compatibility query.

  * ANALYSIS_0001 (postgres/msp.go:531): the Delete cascade's
    `UPDATE tenants SET msp_id = NULL` omitted explicit
    `updated_at = NOW()`, relying on the schema-side
    `sng_set_updated_at` BEFORE UPDATE trigger to fill it in.
    Within the same file, AssignTenant (line ~709) and
    UnassignTenant (lines ~726, ~760) both write `updated_at =
    NOW()` explicitly; the memory backend's matching cascade in
    internal/repository/memory/msp.go also sets `t.UpdatedAt =
    r.s.clock()` explicitly. The Delete cascade was the only
    trigger-dependent path across both backends + both binding-
    mutation neighbours. Long-term fix: write `updated_at =
    NOW()` explicitly so the audit story is uniform and the
    layer no longer silently depends on a schema-side trigger.

  * ANALYSIS_0005 (handler/tenant.go:113-116): the pre-existing
    tenant list handler still used `map[string]any{"items":
    items, "next_cursor": res.NextCursor}` which always
    emitted `"next_cursor": ""` on the terminal page. The
    MSP / alert / baseline / integration handlers had already
    migrated to a typed-envelope struct with
    `json:"next_cursor,omitempty"` so the field is OMITTED
    on the last page, matching the OpenAPI `nullable: true`
    contract. Long-term fix: replace the map with the typed
    envelope to align the surface. Regression test
    TestTenantList_OmitsNextCursorOnTerminalPage decodes the
    body into map[string]any so the absence of the KEY is
    wire-format-true (not just the struct field being zero).

  * ANALYSIS_0003 (handler/msp.go:1157): BulkClaimTokensRequest.
    TTLSeconds was `int`, which is 32 bits on 32-bit platforms
    (max ~2.1B) and 64 bits on 64-bit platforms (max ~9.22e18).
    The maximum decodable TTLSeconds therefore varied by
    architecture even though the MaxClaimTokenTTLSeconds cap is
    the same on both. Long-term fix: pin TTLSeconds to int64
    explicitly so any value the JSON decoder accepts is checked
    against the same cap regardless of where the server runs.
    MaxClaimTokenTTLSeconds also retyped to int64 to keep the
    comparison free of implicit conversion. Server only ships
    64-bit today, so the practical wire behaviour does not
    change; the type is now honest about the bound.

  * ANALYSIS_0004 (rbac/msp.go:152): ListAuthorizedTenants
    bounded by ListAuthorizedTenantsMaxPages (5000) returning
    ErrTooManyMSPBindings on overrun, which writeBulkError maps
    to a 500. There was no logging or metric emission as the
    loop approached the cap, so an MSP growing toward 1M
    bindings hit a sudden 500 cliff with no advance signal.
    Long-term fix: introduce ListAuthorizedTenantsWarnPagesThreshold
    (= 4000 = 80% of the cap) and emit a warn-level slog entry
    exactly once per call when the threshold is crossed. The
    log includes msp_id, user_id, permission, pages_fetched,
    warn_threshold, and hard_cap for operator triage. Once-per-
    call (not once-per-page) so the journal is not spammed.
    Regression test TestListAuthorizedTenants_WarnLogAt80Percent
    uses an infinitePagesMSPRepo stub to drive the loop to
    exhaustion and asserts the warn log is emitted exactly
    once with the expected structured fields.

  * ANALYSIS_0007 (service/tenant/branding.go:210): the
    BrandingResolver.InvalidateAll round-24 fix uses the Go 1.21+
    `clear()` builtin. The reviewer's concern was project-go-version
    compatibility — verified go.mod specifies `go 1.25.0`, so
    `clear()` is fully supported. Acknowledgement-only; no code
    change.

Local verification: go test -race -count=1 -timeout 300s ./...
\u2192 24/24 packages green. go vet ./... clean. gofmt -l on the
touched files clean.

+++ internal/handler/msp_test.go
@@ -3346,7 +3346,7 @@ func TestMSPHandler_BulkClaimTokens_RejectsAboveMaxTTL(t *testing.T) {
 
 	cases := []struct {
 		name string
-		ttl  int
+		ttl  int64
 	}{
 		{"just over the cap", handler.MaxClaimTokenTTLSeconds + 1},
 		{"overflow territory", 1 << 40},
